ingestor/extraction_worker/requirements.txt
requests
google-cloud-pubsub
google-generativeai 